Add HTTP requests and analytics.
As usual, mostly copied, but with modifications. The biggest change is
wrapping the URL transfer object. We used to use some utils from an SDK
example, but it didn't fit all that well and we ended up with a couple
different things. Everything should be able to use the wrapped request
now.