Add plex.tv account handling.
Still pretty basic, but we use information from the registry to initialize
the plex.tv account information, and then validate the token. If all is
well, we publish our device information.

There's still much to do around the MyPlexManager, but most of it needs to
come later. This establishes a separation between the account information
and the request making piece. The latter is likely to inherit from a
Plex Media Server object that doesn't exist yet.