Allow HTTP requests to specify timeouts.
Respect a timeout value on the request context and fire the callback with
an invalid response if the timer lapses before the request comes back. Use
the new timeout to make sure that reachability tests don't take forever
(currently 20 seconds).

Additionally, timers now use Callable for callbacks instead of a special
method name on a listener. And requests that fail to start fire their callback
with an invalid response.