Add basic PIN auth screen.
This is an old style code registration screen, so the screen itself is
trash, but it demonstrates how to do some things on top of PlexNet. The
account can now be signed out of to toggle the welcome screen, and the
PIN screen makes a couple requests to plex.tv and uses a timer. It also
uses an Application event to close the PIN screen only after the token
has been validated, instead of having to assume that the validation
would pass like the old PIN screen.