PlexNet: Various media related objects.
This basically adds the rest of PlexNet: PlexItem, PlexMedia, PlexPart,
PlexStream, PlexTag. They depend on one another and have to be written
from the bottom up, but can't easily be tested until they all exist. To
that end, PlexHub has also been added and a dummy request for hubs is
made to reachable owned servers.