update servers reachability immediately (on first response)
There seems to be no reason to wait until ALL tests have completed to
set the active connection, if a servers current active connection is
invalid. On subsequent reachability responses, we will only reset the
active connection if the current is unreachable, or not local while the
new one is.