Containers, HBoxes, and the beginning of real layout!
Define a basic container concept for components that consist of other
components. Containers are drawn by simply asking all of their children
to draw themselves into their own regions, and then an array of all the
child regions is returned. This allows the compositor to be solely
responsible for rendering the actual regions to the screen.

The first actual container is a horizontal box that is pretty fully featured.
It supports spacing between elements as well as options for whether and how
the container area should be filled if the children don't naturally equal
the full available width.

Containers also introduce the idea of having and needing layout. Before a
container asks its children to draw themselves, it lays them out with
coordinates and dimensions.