Initial work on focusability.
Components can declare themselves focusable, and the component screen tracks
the currently focused item. If there's a focused item when the screen is
drawing itself, a focus box will be shown around the component's borders.

There's still plenty to do. There's nothing in place yet to shift focus from
one element to another in response to remote control events, and there are
likely a variety of small things that are wrong about forcing ourselves to
destroy things and start over.