Allow components to have explicitly set focus siblings.
Add `SetFocusSibling` to components so that callers can explicitly set up
focus chains. Tweak Box to do this automatically for all of its focusable
children. The ComponentsScreen also now handles focus shifts when the focus
chain has been set explicitly.

We will probably want to fall back to some general focus handling when we
don't have anything set, but that is left as a TODO.