Take a stab at a PaddingMixin, especially for Labels.
Add a mixin for setting padding on elements. This is currently only used
on Labels, but the idea is to set some padding that will offset the rendering
area a bit from the actual bounds of the frame. As in HTML/CSS, the
background of the label is still set in the padded area.