Add Deactivate method to the Components screen.
Normally, we need to clear any bitmaps, regions, sprites and fonts when
we push a new screen (deactivate the previous screen) to conserve
memory. This will impede a dialog screen from working, so we'll have to
revisit this and the compositor.reset(), to handle keeping components
when stacking a new screen.