replace the bitmap width the scaled bitmap (if applicable)
Now that we are keeping a reference to the bitmap, we probably want to
replace the bitmap reference with the scaled version. If we do not, we
end up keeping the original bitmap in memory. We could reset the
reference inside imageScaleRegion, but that could have an adverse effect
on the on url placeholders.