Use events to allow components to suggest they should be redrawn.
Not all components will be able to redraw themselves since some components
may be drawn to a composite region instead of to their own region. So instead
of having the image redraw itself when a bitmap comes in, it now triggers an
event to suggest that it should be redrawn. The compositor screen listens to
this event on all components that it drew to their own sprites.

We should probably be calling `Off` somewhere though, so that there aren't
still listeners on components that aren't on screen anymore. A `Destroy`
method has been added to `Component` for this purpose, but nothing calls it.
We probably get away with the current approach though, because the component
ends up with a reference to the listener, but not the other way around. So
if nothing else is referencing the component anymore, it'll get garbage
collected.