Create a Composite class and turn the Card into a basic example.
Create a `Composite` class. A `Composite` is a container that draws all of its
children to its own region instead of letting them draw to their own. This
allows us to build more complicated components out of building blocks, but
get the savings of having a single sprite for the component once it's built.

The `Card` has been turned into an incredibly basic `Composite` as a sort of
proof of concept. The actual card will need additional functionality.