Add HUB component
This is definitely rough. I am not sure if we want to inherit another
class as the base instead of an HBox, but it seemed to fit well
here. We (I) should probably look over some existing apps/source to see
how we decide on hub style, and what image to use (thumb, art,
composite, etc).