change HubClass to inherit from ContainerClass
A lot of this still feels wrong, but it works. Feel free to rip it apart
and correct the wrongs! The basic Image components have been change to
Card components. Do we want to be able to pass the dimensions to the
card, or is setting the frame the correct way?