Allow position and dimensions to be updated independently.
Split SetFrame into its components parts, SetPosition and SetDimensions. For
containers, check to see whether the values being set are actually different.
If the dimensions have changed, then note that we need to reperform layout.
If only the position has changed, then the container can update its own
position and that of all its children without redoing layout.

This often allows slightly complicated layouts to avoid multiple complicated
layout calls. In particular, consider a horizontal box made up of hubs. The
box wants to know the preferred width of the hubs, which requires the hubs to
do a measurement/layout pass. Knowing the preferred widths of its children
allows the box to do its own layout, at which point it calls `SetFrame` on
its children. The hubs now notice that their dimensions are the same and only
the position has changed, so they can skip another layout pass and just
update the positions of their children.