add a rudimentary header to the HomeTest screen
How do we want to handle the header? This code completely temporary. I
just wanted to get some buttons above the hubs to see how it
interacts. It seems like we might need a createContainer or something to
add objects to a parent without the parent resetting dimensions. I could
be missing something, as I've spent my time working on the manual focus
/ calculating distance between components.