performance: reuse a region if applicable (compInitRegion)
during a component draw, we call InitRegion() which will create a new
bitmap/region every time. This will allow us to reuse the region if we
already have one by simply clearing the region with the expected
bgColor. The component will still redraw what is needed, but we no
longer waiting on a roku to generate a new bitmap/region.