performance: Image comp makeCopy=false as default
We normally do not need to make a copy of the bitmap when setBitmap is
called. An Image components are initialized with a unique region, which
we'll reuse and draw the bitmap into. The only caveat (as of now) is
that composites with image components will invalidate the image
region to conserve memory. This is ok because we draw the bitmap into
the unique composite region, leaving the URL bitmap is untouched.