temporary Image Transcoding for the Image Class (TODO PlexNet)
This is temporary as we don't have proper image transcoding available
yet. For now, if we expect an image to be transcoded, we can pass the
image source as an AA (container a url string and server object). When
the image is ready to be drawn, it will convert the image url to the
proper transcoded url. We cannot pass the transcoded url prior to the
image component because the dimensions can be dynamic.