Rework font registry into a wrapped singleton.
Instead of attaching the font registry to the Application, we can use a
singleton. Wrapping the actual roFontRegistry also allows us to do a couple
nice things like hide HD/SD size differences from the caller and hide the
font name from the caller. At least for now, perhaps some day we'll need more
than a single font for text and another for icons, but I hope not.