Initial work on a ComponentsScreen.
The idea here is to have components that know how to draw themselves
within their own bounds and then a screen composed of these components
that draws them to the actual roScreen.

While we're playing with this idea, the screen is forced to a component
test screen. Currently the only defined component is a colored block, and
the test screen simply shows a single block at a fixed position.