Add DropDown component
The drop down component will maintain an array of AA's to build the
required components when toggled. As of now, all components are assumed
to be buttons. It will also keep track of the components it built inside
another array, and won't modify the screen components list. More cleanup
and features are definitely on the list as the come ( vertical scrolling
will be a fun one )