add a new container for manually created components
We have the need to create components on the fly, after we have created
a components screen. Things can get messy if we modify the base
components container. Instead we will assign a new key to the
componentsManual container to keep track of manually created
components. E.G. the description on a home screen or section screen will
need need to be displayed or removed based on the focused content.