CancelRequests was broken in the Application singleton
I am assuming the context.callbackCtx was the callbacks context. It was
not set in the during the httpCreateRequestContext routine, but is
now. I have also added an option to set the context.screenId, in case we
might be using context that isn't a screen.

CancelRequests was also broken as it tried to call the cancel method on
the wrong object. It's now one more layer deep.