do NOT show a screen if it's not the active screen
This might seem a bit odd, but this happened previously due to the
CancelRequest feature being broken when pushing/popping a screen. It
seems like a good measure to add to any screen, but especially ones that
might need to request data before being able to process and display it.