use the loading modals when pushing or popping screens (on a timer)
On every screen push or pop, we will create a loading modal that will
present itself if 500 ms has passed before a components screen has drawn
it's first object ( we call Application().CloseLoadingModal() right
after we draw the components ). The timer will also call the Deactivate
or Destroy method on the previous screen when the modal has been
canceled/closed, created or a new modal has been created. We need to do
this so we can draw the modal with the existing components, otherwise
we'll just get a blank canvas with a loading modal.