support AlignmentMixin in BOX containers
This will allow components to use their specified height or width in a
BOX (VBox/Hbox) container. Components will be placed at the same x or y
offset as they were previously, but one can override this by setting a
parent alignment on the components (phalign or pvalign). If the parent
alignment is not set on the component, the component alignment is used,
which allows this change to not adversely affect the placement.

E.G. VBox with 6 button components. The 6th button set to width: 100px,
halign: JUSTIFY_CENTER (button default), and phalign: JUSTIFY_RIGHT.

When drawn, all buttons will be set at 200px and centered, except for
the 6th button. The container will now adhere to the dimensions and
placement, rendering the 6th button as 100px width, justified to the
right of the box.