remove demandLeft from base shifting properties
Each parent will need to include their wanted left offset when any
element is focused (demandLeft). Hubs are the only example for now that
work this way.

e.g. Any focus on the first hubs components will being the entire hub to
the left of the screen. All other hubs will be brought to a specific
offset (300px) until we reach the end.