grid screen: rework lazy loading and grid chunks (per model)
The Roku 3 is the only model right now that can handle processing
background events without much, if any lag on the UI. We have to handle
other Rokus a bit different: load less data up front and less data on
the fly until scrolling has stopped.

The major time spent is updating the offsets for the off screen
component (testing with a section having > 1500 items). Hopefully we can
find a quicker way to update the components offsets. The PoC used the
"parent" as a quicker way, but as of now, we are not updating the
parents offsets. Updating ONLY the parents, would be something to fiddle
with.