encourage memory cleanup during screen deactivation (GC)
We are pretty blind to the available memory. If we end up hitting a
limit, the Roku will just reboot. We can use r2d2_bitmaps on port 8080
to see the used/available memory for bitmaps, fonts, and other
downloads, but it doesn't give any insight to the memory used for
anything else within the channel.

I can't see any way not to use this, unless we start using some sort of
lightweight reference for the virtual screen stacks.