Initial pass at refocus (global)
There are still many tweaks to optimize this along with fixing the
Components screen shifting methods to handle this correctly.

It would be optimal if we could ignore a full reload on activation, but
that won't be possible due to memory constraints. We'll probably need to
look into keeping some lightweight m.components counterpart to be able
to redraw the screen without querying the PMS. Which is most likely a
requirement because we'll need to cache the data for some amount of time
due to the dynamic nature of the hubs. Caching will introduce other
issues like caching the watched state, which will need to update and
display.