option to ignore parents onScreen status when shifting
This is prep work for the preplay context screen. The screen will need a
"grid", but we do not need to over complicate it with all the grid
logic. It just needs to be a simple hbox, utilizing the default shifting
methods, except for the fact we'll only be shifting ONE item at a time
instead of the entire parent.