add method to destroy and clear components
This will be used to replace any call to m.components.clear(). It's not
safe to just clear the AA as some references are not cleared from
memory. I.E. Sprites... the reference will be lost but it will still be
attached to the compositor with no way to lookup/clear the sprite later.