support container layout for unknown image dimensions
There are times we have no idea what the dimensions of images will be
before we request them. We do however, have the need place them in a
specific layout. This will allow us to specify the wanted width and
height of the container, and place the items where they need to be
according to the layout.