handle processing urlEvents and textureEvents outside of the main loop
There are some times we might animate (shift sprites around) which
would block the message loop. In this case we can include
Application().ProcessNonBlocking() in the blocking loop to handle any
image requests. E.G. a animated progress bar will be added to the
preplay screen. This progress bar will probably be displayed before the
textures come through. During animation loop, that is blocking, we will
want to process and display these images instead of waiting for the loop
to complete.