rudimentary support for unobtrusive overlays on cards
This is a little tricky as we calculate the image size, grid size, hub
size and others based upon the available space given. This means, for
now, we'll need to calculate and override specific dimensions while we
perform the layout.

The next step is to obtain a blurred bitmap of the poster for the
overlay background. We'll see how much more strain this puts on the UI.