Move image transcoding into plexnet, fix home screen.
Add image transcoding functions to plexnet (PlexObject and PlexServer).
This also fixes converting URLs to loopback, which fixes shared thumbs.
So far only the home (hubs) screen has been completely updated to use
the new methods correctly though. This is because I think we need to
tweak our Image component. It has been keeping an AA of data about the
image source, but I think what it really wants is a reference to the
PlexObject. That way it can ask the object to build a transcode URL
for an appropriate attribute instead of trying to precompute all the
attributes that may be needed.