Initial implementation for server selection. #124
Try to choose the best server possible after startup. This is somewhat tricky.
For starters, we may have servers remembered from past runs, servers from
plex.tv, and servers from GDM. And some of those are coming in during our
search. But we can't wait for everything, we want to select something as soon
as possible without choosing the wrong thing. Our basic approach is:

- Evaluate the state of things every time we get a reachability result.
- If we found the server we hoped to find (the last used server), we're done.
- Otherwise, see if this server is the best we've seen so far
  (preferred > owned > local > remote shared)
- See if we're still waiting to hear back from anything better than our best
  option so far. So if we have no preferred server, the first owned server
  is good enough. If we have no owned servers, the first shared server is
  good enough. If we find a shared server first, we have to wait for the
  owned tests to finish.

There are a few things pending still.

- We're not remembering the last used server, so the preferred server stuff
  is speculative.
- We need to keep track of whether or not we're waiting for a response from
  plex.tv/resources
- If you have unavailable owned servers, you may have to wait 20 seconds for
  the connection tests to fully fail. That shouldn't be too common, but it
  feels pretty slow.