Tweak video player to make it a singleton.
We want both a long-lived singleton VideoPlayer that can be used by the rest
of the app to interact with video playback (e.g. remote controller) and manage
playback of multiple items (e.g. multipart, or play queues). But we also
need something to look like a screen and do things like handle messages. The
lines are pretty blurry, so for now, we have a single object playing both
roles.