add support to focus components manually on an overlay
Previously we had to always set `FocusNonSiblings=false` on any
focusable item on an overlay. This however was not ideal as all
components on the overlay were required to be in the same parent, or we
had to manually set focus siblings.

A pinPrompt (coming soon), is a good example when we want to support
manually focusable components.

The settings overlay is an example when we still require
FocusNonSiblings, as we need to keep the focusable components contained
to the parent. e.g. setting RHS scrollable vbox.