Allow overlays to be layered without a stack overflow.
We were replacing the OnKeyRelease function twice such that hitting a d-pad
key on the second overlay would infinitely recurse and overflow the stack.
Whether or not we _should_ show multiple overlays is another matter, and this
change doesn't do anything about the fact that the screen dims twice making
it totally obscured.