Tweak loading timers.
Our GDM timer was set to 10 seconds, and connection tests are set to 10
seconds, which means that things could take 20 seconds to come back. But
our timer was only allowing 15 seconds. (Incidentally, this means we'd
give up on the timer and show the failure dialog, and then get the event
and show it again. A separate problem.) 10 seconds is probably excessive
for GDM though, so I've turned that down to 5 and just added a couple seconds
of buffer to the safety valve timer.