Add really simple support for per-user prefs.
This just allows getting and setting per-user prefs. We're not doing any
sort of fallback to shared prefs (which I don't think we want to) or trying
to initialize a user's prefs when switching to the user for the first time
(probably want to) or trying to clean up after signing out as a user (not sure).