Wire settings dialog up with AppSettings.
Define all preferences completely in one place, including the default value
and things like whether or not restricted users have a forced value. This
should make the callers much simpler (e.g. `GetPreference("local_quality")`
instead of `GetPreference("local_quality", "7")` to specify a default value).

The settings dialog gets its values from the registry now and writes changes
to the registry. It also triggers events for settings changes (which is
probably wrong, that's probably the responsibility of AppSettings). As a PoC,
the logger listens for changes to the log level and acts accordingly.