fix crash due to triggers order of execution #173
Regardless of Plex Account status, we'll need to run the
PlexServerManager singleton to wire up the `change:user` trigger in the
right order. We just need to make sure we create the PSM trigger for
`change:user` before the Application adds one, so we can place
`PlexServerManager()` elsewhere if needed.