add support to override the GridScreen path #180
This will allow us to manually add a pivot on the (episode) PrePlay
screen for the specific season (Go to season #). An episode preplay
screen we will not have a valid plexObject for the season without
creating some sort of synthetic plexObject, or executing another api
call to the PMS. Instead of faking a plexObject, or adding more api
calls, we can simply override the wanted path (key) for the GridScreen.