add support to override the PrePlayContext screens path #180
This will allow us to manually add a pivot on the (episode) PrePlay
screen for show (Go to show). An episode preplay screen we will not have
a valid plexObject for the show without creating some sort of synthetic
plexObject, or executing another api call to the PMS. Instead of faking
a plexObject, or adding more api calls, we can simply override the
wanted path (key) for the PrePlayContext screen.