support remote play button for shows/seasons (play On Deck) #179
We do not have play queue support (yet), so the next best thing is to
play the On Deck episode when the user presses the remote play button on
a show or season.

Seasons to not have API support for `includeOnDeck=1`, however already
query the shows key when pressing play from a PrePlayContext screen. We
may have to add extra support to query the show from a seasons item if
we ever display seasons in any other screen than PrePlayContext.