Fix memory issues (internal scaling) - load less, unload more.
The texture manager used to handle scaling images, so the bitmaps were
always reusable (the size requested was always returned by stretching
the bitmap).  we'll need to be a little more aggressive with unloading
components, now that we need the ability to scale the image while
maintaining aspect ratios (scale-to-fill). If the requested image
doesn't match the texture manager dimensions, then we'll have two
bitmaps using memory.

We may want to look into unloading the texture manager url during
SetBitmap() - TextureManager().RemoveTexture(m.source). It would remove
the image from the cache, which could affect composites performance on
any redraw.