Initial Play Queue implementation and audio player integration.
Define a Play Queue object and always use that in the audio player instead
of setting a more manual context of tracks. The play queue supports the
notion of updating itself, and the audio player registers for change
notifications and updates its context accordingly, but nothing is actually
triggering a PQ update.

The album screen has been tweaked to create a play queue in a limited way.