Advertise play queue support for remote control.
Advertise that we're play queue compliant in the remote control protocol,
augment our timeline data accordingly, and handle the refresh play queue
request.

While I was in there, I discovered that we crashed if web tried to connect
to us. The crash(es) have been fixed, but it still doesn't seem to be working.