Mostly fix web remote control.
This fixes a crash I reintroduced and improves the container key for play
queues. It's still not perfect though because web doesn't seem to tell us to
refresh the play queue ever. I'm not sure if that's purely a web bug or if
we're supposed to periodically refresh the play queue even when nobody has
told us that it might have changed.