Reimagine Triggers as a mixin.
Convert the triggers stuff to a mixin and add it to both Component and
ComponentsScreen. This should make it easy for components to register
listeners and have them cleaned up automatically.

The header (mini player) and album screen have been modified to listen to
the audio player in this way.