Try to always locate focus regardless of current focus.
This shouldn't affect the current UI at all. The update will allow us to
hide focus on a screen, and reuse the last focus item when we need to
show focus again. e.g. Now Playing screen hides focus after an arbitrary
timeout, but needs the ability to refocus the last focused item
later. We can simply set the lastFocusedItem to the current focus, and
invalidate the focusedItem.

This also reimplements setting the focal point correctly, which I ended
up overriding a while ago.