Add a toggle timer to reenable focus for a better UX. #202
Toggle type buttons will change color based on their state. If we set
the button to a disabled state (repeat = false), then we change the
color, but then it looks like we are no longer focused on anything. This
change will show the button being toggled, and then reset the color to
the focus color (OrangeLight), if the button is still the current focus.