Update VBox scrolling logic (relative height)
This will hopefully clear up some confusion when setting the "scroll"
height on scrollable VBox. Previously, the height required the yOffset
to be included. Now we only have to set the height we want relative to
the positioning of the VBox.

Secondary to this fix, is scrolling to the correct position on refocus
when the lastShiftInView flag is set.