Add support for VBox scroll topics/groups.
This is very rudimentary (and probably not named well). It will allow us
to specify a "topic" item to any group of items in a VBox. If any item
in the group is focused, then it will try to position the topic item to
the top, only if the topic item is out of view AND if the topic item
doesn't shift the current item outside the viewable area.