Add support for "continuous" play queues from certain hubs. #230
When playing an item directly from an On Deck or Recent hub, if that item
is an episode then create a "continuous" play queue that will play all
unwatched episodes starting from that point.

Some of the logic here is a bit suspect. It doesn't seem like any clients
are doing this correctly right now, so we're hoping for some clarity.