Rework audio stream language coercion.
First of all, there was a typo that made the whole thing worthless. Besides
that, we were doing some complicated tracking of languages that we'd seen
and failing to choose correctly on files such as [5.1, 2.0, 2.0] all in the
same language. Instead, we just pay attention to tracks in the same language
as what we're going to try to force.