Close all overlay screens on push and pop.
This fixes an issue where the overlays (just the bitmaps) could persist
forever on the screens. It was reproducible in the following scenario:

* Start playing music (now playing screen is displayed)
* Toggle the queue list (overlay is created on the now playing screen)
* Start playing music through a remote, Plex Companion.
* The now playing screen would be closed (pop screen)
* The overlay (queue list) is lost in space... persists on the screen.

This will also fix a secondary issue, as we were only closing the last
overlay when pushing new screens. It's possible a screen could have
multiple overlays, so we need to close them all.