Add cache support to the TextureManager. #178
This will add rudimentary support to cache images for the current
screen. We'll have to update this if we want/need images persisting
across multiple screens. Add `cache=true` to any image you create, and
we will reuse the cached texture as long as the next request matches the
URL and dimensions of the cached image.

Fixes #178