Add playback failure dialog screen.
These errors including the title and text were ported from the official
channel. We will probably want to review the title/text before accepting
this.

TODO: we need to figure out the right method to check for
accessibility/availability of items in a play queue, as checkFiles does
not work.