Defer checkFiles until failure. #254
checkFiles will supply us with the accessibility and availability info
about the current item. It does have the possibility of unnecessarily
delaying the UI, so we should only include it for specific needs. Also
note, the media.choice does check for the availability and
accessibility, however it's current state is a bit broken because
checkFiles is not supported in playQueue items, so they will always
technically be available and accessible. (TODOs added)