Remove the focusNonSiblings logic (deprecated).
We should always use the overlay screen, which was added after the crazy
non siblings focus checks. Basically the focusNonSiblings was added to
ignore focusing to other screen objects that were not part of a
container, e.g. a drop down. We can now ignore this logic, since the
overlay will only focus to the components within the overlay, ignore any
other components on screen.