Support deleting screens in the middle of the stack. #271
It's possible we may push a screen before we have successfully popped
it, or in reality, it's possible the roku may close a screen from
underneath us. The latter is real as the VideoPlayer will be closed if
we show a screen over the top (lock screen). In this case, lets just
delete the screen and cleanup.