Add support to clear textures by screen or overlay.
The now playing queue overlay will contain album images per track in
mixed content. Overlays are attached to the screen, so when we close an
overlay, we need to be able to release the textures. Previously the
textures would only be cleared if the screen was deactivated in any
form. This is still true, however, we will now clear any textures in an
overlay when it's closed. The manually cached items (cacheList) will
also only clear the cache specific to the screen and/or overlay.