Try and process audio feedback immediately.
Just do a quick peek at the message port to see if any audio is waiting
before we process anything else. I still feel that we need a setting to
disable these until we can query the global roku settings.

All I can figure for the dealy is that is could just be another roku
bug. The delays don't seem too frequent, but probably enough to be
annoying. The odd thing, is that the delay seems to happen, when when
the screen obviously shows it can update faster. e.g. moving around a
single hub, zero shifting. I can see the focus border move before the
feedback. The feedback trigger is executed before we do anything...