Stash a reference to the texture request and cancel when applicable.
It was assumed we stashed a reference to the texture request, as we try
to cancel textures when we unload components. I guess the reference was
never created. The reference has been added along with the additional
step of canceling a pending texture, if it exists, while requesting a
texture for said component.

tmCancelTexture has been updated to reflect it's OK to blindly send the
textureRequest context. No need to double check if the context is valid.