Fix action bar focus when MiniPlayer is visible. #282
The issue is outlined in #282. This is more of a shortcut to only fix
the this specific issue when the MiniPlayer is visible. It's possible we
could run into similar future issues. We'll probably cover/fix this
issue in the future when we rewrite the focus logic, but included
below (gist) is another possible short term idea, that may help other
similar issues.

https://gist.github.com/ljunkie/0ddc14b88b13572eb6d4