Rework our cached images in the texture manager.
This will add support to utilize the cached images on the next screen
before we kill clear them. We'll need to update this logic if we need to
keep cache around longer than the next screen.

The user avatar in the header, as well as the background images will
benefit from this, once we start caching them.