Update image fade-in logic (crossfade support)
Previously we were fading in from a transparent state, regardless of the
previous screens image. It probably wasn't noticeable on many screen,
except for the Now Playing screen, until we added the preplay context
switching via the fwd/rev buttons. We will now try and use the previous
image to crossfade with while fading in.