Ignore resetting Now Playing images for the same item.
We already request the images for the requested track when the screen is
created, and we almost instantly receive a trigger from the audio player
with the current playing track. We can safely ignore the trigger if the
item playing is the same one we initially used when creating the
screen. Same logic applies for repeat.