Initial implementation using fwd/rev button on grid screens. #256
This might be more of a POC. I'm not entirely with the scrolling, but
that may be for another GHI (acceleration). I also had to modify the msg
port peek hack as it didn't work at all while scrolling the grid with
the key held. We have many pending events in the queue, so the grid
would keep scrolling even though we had a pending key release waiting in
the queue. Basically, if you held a key down for a short duration, the
grid continued to scroll until it reached the end.