Fix clearing cached images when refreshing the Now Playing queue.
I had initially cleared the cache inside the overlay components `show`
method, however that's normally wrong. We would usually want to keep a
cache for items still on the underlying screen. The Now Playing screen
and overlay are slightly different. On an overlay refresh, we'll need to
add all the existing cached items to a pending delete list, and clear
the unused items after the update.