Initial attempt at Playlist support.
This still needs a lot of cleanup as well as modifications to make it
look more like the invision mocks for audio and video.

There is an issue right now with scrolling performance of a VBox. We
haven't experienced it yet because most of our scrollable vbox
containers have limited children. For now, I have limited the playlist
items in the list to 50. What I don't know is how we want to handle
this. Do we need to show more than 50? If so we'll need to pull some
more logic out of the horizontal shifting because it can cope just fine
with thousands of items.