Only simulate key presses if there's no pending release event.
This reworks how we check to see if we should simulate a key press while a
key is being held. Instead of ignoring timers while a key is held, we keep
a queue of messages that have been received on the port. Before simulating
a key press, we check for any key release events in the queue.