Fix possible crash on Now Playing screen. Fixes #328
We do not want or need to reset the screens onscreen components if we
are currently in the context of an overlay. The overlay handles it's own
focus logic. We do need to rest the components if the vbox is part of a
standard screen, as we may load/unload components while scrolling.