Fix selected item for large play queues. Fixes #331
We typically know what the selected item in a play queue is, and we update
it as we move from one item to another. But we also need to refresh the PQ
from PMS periodically, and PMS lets us know what it thinks the selected item
is. We may not agree, particularly if we just switched items and PMS hasn't
received a timeline yet to update its state. So we generally trust ourselves
more than PMS.

We were already doing that, but one important exception was when the PQ version
changes. This will happen if you shuffle, and that results in a completely new
set of PQ IDs, so we decided to trust PMS implicitly if the version changed. It
turns out that the PQ version can change even when the IDs haven't changed
though, and that was potentially messing everything up. This is now fixed in
one and a half ways (the first change solves the problem, the second would
just make it less likely, but it also seems like a worthy change):

- Instead of trusting PMS implicitly whenever the version changes, we always
  trust ourself if we can find a matching ID. If we can't find a matching ID
  then we use PMS' instead.
- Instead of immediately refreshing the PQ when the item changes, we wait 5
  seconds. This makes it much more likely that PMS will have received a
  timeline and knows what the current item actually is. This should lead to
  more accurate IDs (which isn't a huge deal since we ignore it) and a more
  accurate window in general.