Support shared regions in the block component.
This is rudimentary and I'm not sure how well it will work for
everything, but for now, we need the ability to use shared regions for
item separators. We currently create a region to be shared, but if it's
ever unloaded or redrawn, that link with break. The refactor of shared
separators will be coming shortly.