Refactor track separators. Reusable regions and scrolling anomaly fixed.
We had previously been adding a 1 px separator to the vbox a track is
part of. This would end up shifting the vbox by 1px extra every time,
which we account for, however when you scrolled up, you'd noticed an
awkward small shift when reaching the first item. The 1px separator is
now part of the component itself, so all of our vbox items are once
again homogeneous.